fix: separate compare and check benchmarks
