+++ Cargo.toml
@@ -35,6 +35,10 @@ panic = 'abort'
 name = "check"
 harness = false
 
+[[bench]]
+name = "compare"
+harness = false
+
 [lib]
 bench = false
