+++ benches/check.rs
@@ -24,26 +24,5 @@ pub fn check_benchmark(c: &mut Criterion) {
     });
 }
 
-pub fn compare_benchmark(c: &mut Criterion) {
-    let temp = tempdir().expect("create tempdir");
-    let path = temp.into_path();
-
-    let current_dir = env::current_dir().expect("get current dir");
-    let app = dotenv_linter::cli::new(current_dir.as_os_str());
-    let matches = app.get_matches_from(vec!["dotenv-linter", "compare", ".env", ".env.compare"]);
-
-    fs::copy("benches/fixtures/simple.env", path.join(".env")).expect("copy .env file");
-    fs::copy("benches/fixtures/compare.env", path.join(".env.compare"))
-        .expect("copy .env.compare file");
-
-    c.bench_function("dotenv_linter compare", |b| {
-        // Disable output to STDOUT
-        #[cfg(not(windows))]
-        let _print_gag = Gag::stdout().expect("disable stdout");
-
-        b.iter(|| dotenv_linter::compare(black_box(&matches), black_box(&path)))
-    });
-}
-
-criterion_group!(benches, check_benchmark, compare_benchmark);
+criterion_group!(benches, check_benchmark);
 criterion_main!(benches);
